requirements.txt
requests
aiohttp[speedups]
python-dotenv
pydantic
